Add websocket based messaging
We import an attempt for server-client communication via websockets from
an old prototype into exceed.

I questioned whether I should use the whole Spring Websocket / sockjs
shebang, but in the end decided against it. Maybe it will prove an
mistake in the future, but I want a quick internal way of working with
messages, squeezing out a bit of perceived performance over JSON.

For now there is no fallback in case the browser does not support
websockets. If we actually need this, the plan is to use a simple
two-tier system and not bother with half-measures like other streaming
options.

Either the client supports websocket or it has to live with using AJAX
posts and long-polling for server to client communication (for now just
reloads).

Exceed Messaging
================

The messaging systems works by exchanging JSON objects that are parsed
into typed Java classes on the server. We use svenson JSON lookahead parsing
as usual.

Message Handler
---------------

To receive messages from the client, one needs to implement the
EditorMessageHandler interface on a declared spring bean.

public interface EditorMessageHandler<T extends IncomingMessage>
{
    void handle(MessageContext context, T msg) throws Exception;

    /**
     * Returns the message type handled by this handler.
     *
     * @return  message type
     */
    Class<T> getMessageType();
}

The getMessageType determines the type of messages handled by the
message handler. Each message type corresponds with type name

	"message.MyMessage"

where "message." is a fixed prefix and MyMessage is the simple class
name of a message class declared by the getMessageType method of a
EditorMessageHandler. Each handler must use a unique type name.

The client can use the hub module to send requests.

	hub.send({
		type: "message.MyMessage",
		param: "abc"
	})

The client has to define a type attribute with an existing message type
name and all other properties needed by the message in question.

The websocket implementation will automatically convert these JSON
objects to the correct type and invoke the EditorMessageHandler with the
message and a message context that provides access to the larger context
of the message like the runtime application it is happening in etc.

There are three base classes for messages:

de.quinscape.exceed.message.Message

Most general base class providing only the correct type attribute. Is
mainly needed for messages that go from the client to the server for
identification and routing purposes on the server, in some cases it is
however useful to also extend it for messages that go from the server to
the client which need a type identification method.

de.quinscape.exceed.message.IncomingMessage

This is the base class for all messages that go from the client to the
server. It provides a MessageMeta attribute which provides basic
information about the message and is filled in automatically by
the hub module.

de.quinscape.exceed.message.Query

This is the base class for all messages that go from the client to the
server where the client expects a response to the message.

Query / Reply
-------------

Asynchronous messages can be really nice, but for a large variety of use
cases we need something very similar to the classic request/response
mechanism.

Messages that extend the Query base class can be replied to by calling

	ctx.reply(msg, response)

on the MessageContext provided along with the query message.

The response object can be any object that can be converted into JSON.

The client side will automatically wrap the whole request/response pair
in a promise so you can just write

	hub.request({ type: "message.MyQuery"})
		.then(function(data)
		{
			...
		}

to receive the response data from the server. Don't forget to .catch all
promises that you don't return from your methods.