Views: Refactor to render function rendering
It suddenly occured to me that I've been doing this dynamically
generting views thing wrong.

One problem that remained was that React wouldn't accept a changed view
as the same view component, which is of course totally right in a way
since it's a totally new component definition.

This turns every view update with a changed view to a total DOM
recreation and loss of all component states.

But it's totally unescessary because we don't really need to create new
view components. We just need to have one view component that passes on
the rendering to a dynamically created part.

We already did go the way of rendering the JS source code for a js
Function. This function used to be a component creator.

Now we just change the function into render function that accepts one
component parameter. We use closures to make sure we bind all arguments
like they were and just replace the component.

Then we refactor our viewCache into a render function cache and wrap the
render functions in an JSON object with the keys "fn" (our function) and
"version" (what used to be ViewComponent.version).

We refactor the current view render method into a general promise based
element render method. We use a "decorateView" function to decorate the
one view component to the structure we need for our views, currently the
view, the in-page editor and the throbber component.

This change enables incremental rendering in view-change cases and also
allows components to keep their state on view-changes, provided they
still have the same location or a ref prop.