Build: Introduce "exceed-services"
For all source code in src/main/js of the main project, dependencies are
really easy. We just require() them and browserify bundles them up into
our main bundle.

We will have to have the ability to add parts from the projects using
exceed, be it custom components and/or actions and other things.

For this purpose we make browserify provide a an external export for
"exceed-services" which will contain references to all services we want
to make available to the general browser environment / project bundles.

The module that will be returned as "exceed-services" from the outside
is src/main/js/services.js in the main project.