Refactor Datalist
The old datalist design might was tied too closely to the mechanics of
the query mechanism which is bad for a general data exchange format.

The on the fly generation during JSONification was also overly hacky and
not congruent with finally having to reparse the (partial) data
structure when it comes back from the client.

We implement a new data list class that uses normal domain type models
as "types" entries and has a very simple ColumnDescriptor class for the
"columns" property.

We change the query mechanism to create a datalist copy out of its inner
classes.

The "rows" array property filled with the actual domain data gets copied
into maps and list and receives converted object properties.

We extend the expressiveness of our type system a bit to support the
following scenarios:

	* simple datalist of domain types
	* datalist of joined domain types
	* Properties containing Maps or Lists of other domain types or
	  property types.

The latter will be mostly used for internal model-editing but might end
up in support for JSON db fields, too.