Implement CodeEditor
Implement a ace code editor that edits an XML-variant of our view model
by converting it on-the-fly to XML when loading the model into the
editor, then allowing the user to edit it and finally convert it back to
JSON before previewing oder storing it.

It might seem a bit odd to suddenly use XML for that, but XML is not
that bad for limited use. It allows a relative noise-free, commonly
known data-format that is far less noisy than the JSON format. The JSON
format is also a bit unwieldy to edit with all it's sub-structures and
even harder to autocomplete.

JSON is a good transport protocol for Js applications though. We could
theoretically use XML and an XML parser to parse the views, but that
would also mean turning them into a far less lightweight DOM and expose
us to a lot of crossbrowser inconsistencies.

Our expressions as "{}" wrapped strings fit just nicely into XML.

Component-Descriptor
--------------------

We implement a first general environment for data driven autocompletion.
The view is very special among all the models. While have some of the
structural complexity, none of them have the need wrap elements in one
another with special rules about which element goes into which under
which conditions.

The other models are either really simple structures like the domain
model or are more homogenous in their structures like processes our
routing tables.

This leads to the decision of having the special code editor just for
views alone and have specialized UIs for the other models.

There is also the complication of the dynamic and data-driven nature of
components.

Components can define rules for the kind of children they generally
allow, they can define classes to classify themselves and have parent
rules that validate the existence of parents with certain classes.

{
    "components": {
        "DataGrid": {

            "classes" : [ "element" ],

            "propTypes": {
                "type": {
                    "client": false,
                    "type" : "QUERY_EXPRESSION",
                    "rule" : "domainType()"
                }
            },
            "propWizards" : {
                "type": {
                    "wizard": "DataGridTypeWizard",
                    "title": "Configure Query",
                    "description": "Open Query join dialog for complex queries"
                }
            },
            "childRule" : "component('DataGrid.Column')"
        },
        "DataGrid.Column": {
            "propTypes": {
                "name": {
                    "rule": "fieldOf(parent.prop('type'))",
                    "required" : true
                }
            },
            "templates" : [{
                "wizard" : "DataGridColumnWizard",
                "title" : "configure..."

            }],
            "childRule" : "hasClass('element')"
        }
    }
}

Here we see an extract from the Datagrid component.json as far as it
refers to component rules and completion.

The DataGrid component classifies itself as "element" which is a class
we introduce with the meaning "this is a general element that can be
used whereever 'normal' elements are used". There will be a lot of
components with that class, but not all of them.

For example, DataGrid.Column components are only valid within datagrids,
so they're not class "element".

The "childRule" property is an expression expressing what kind of
children are allowed for a component. Here we simply define that data
grid only excepts DataGrid.Column children. DataGrid.Column in turn then

allows all "element" children.

Completion
----------

We use the extensible nature of the ace editor to provide an extensible
data-driven way of influencing autocompletion of component, prop names
and values.

Compoment completions derive from the aforementioned child and parent
rules.

Prop names are completed from their definitions inside the component
descriptor.

Prop values are completed by evaluating the "rule" property of property
declarations.

The type attribute of DataGrid is defined as "domainType()" which
corresponds to a context less operation in the new
PropCompleteEnvironment that produces a list of AceCompletion objects.
These objects are a mix of the ace editor completion conventions and
some additional properties for handling them in our system.

The domainType() operation simply accesses the list of existing domain
types in the current application and produces an AceCompletion for each.

Rule expressions can also refer to expressions in the component props of
other components

The name property of DataGrid.Column is such an an expression

	fieldOf(parent.prop('type'))

The parent identifier delivers refers to the parent component of the
component in which we are auto-completing. The type prop of the Datagrid
contains a partial QueryDefinition expression, the part that defines the
QueryDomainType itself to be exact.

The .prop() operation can acess the type expression of the DataGrid and
can identify it as query expression by the type: QUERY_EXPRESSION
annotation. The .prop() operation extracts the QueryDomainType
definition from the expression and feeds it to the fieldOf operation
that lists are fields that are not yet used.

Templates and Template Wizards
------------------------------

Each component can define an array of component templates which are
either a component model in JSON form that will be inserted as
completion or a wizard.

Wizards
-------

In addition to this declarative way of defining completions, there's
also the possibility to define so-called Wizards which come in two
types (producing either components or prop values) and two flavors:

	* React Component
	* JavaScript function

E.g. the template wizards are functions that can produce content by invoking
editor functions.

DataGrid.Column defines such a template wizard. The wizard just inserts
<DataGrid.Column name=""/> defining the only (and required) attribute.

Then it navigates between the quotes of the name prop and triggers
another auto-completion process.

So when the user select that wizard from the context menu, the user will
be instantly provided with a choice of field names to autcomplete.

Prop Wizards
------------

In addition to this declarative way of defining completions, there's
also the possibility to define so-called Wizards which come in two
flavors:

  * React component
  * functions

React component wizards are displayed in a module dialog and provided
with a prop that lets them return the completion after the user has
configured it via dialog.

functions are just a piece of js code that uses ace editor functions to
do its job.

Preview
-------

The editor also has a live-preview function. As the user edits the XML
document, the editor keeps converting the current XML state back to a
JSON model and then POSTs that to the server with a special preview
header.

In preview mode, the rendered view does not use the view model as it is
on disk and loaded into the server-side application but instead
on-the-fly generates a view and view-data just for the current editor state.