Refactor RuntimeApplication
Refactor RuntimeApplication for better testability.

We rename the old implementation to DefaultRuntimeApplication and
extract and interface containing all the model and domain related
methods that can be used in a number of contexts as information source.

We keep all the nitty gritty resource loading and signalling and CSS
collecting stuff inside the default implementation and expose that to
a number of controllers.

This is not as clean as it could be but it works for now and increases
testability for a lot of completion query cases and similar. Including
cases where I already evaded the issue by only providing the domain
service as context when the application was the natural context (prop
completion)

We add a TestApplication and a TestApplicationBuilder that helps
creating RuntimeApplication instances for tests.

We clean up converter resolution a bit to not rely on Spring but on a
new DefaultPropertyConverters class we can also use in the tests.

Should still be easily replacable if needed.