Refactor Routing
Move the resolve method to the RoutingTable itself and make it actually
extract path variables.

Preextract the var name from variable nodes and handle isVariable based
on that.

Add a process name (was missing) and an alias property to mapping.

Mappings need to either point to a view or a process. The mappings need
a unique name which is take from either the view or the process or the
alias property.

So by default, mappings are referred to by the name of their target
which is then assumed to be only mapped one time obviously. If a view or
process needs to be present multiple times, the mapping needs to define
an alias.