Expressions: Add array literal, but not assignment
We were still missing an array literal.

I initially also added an assignment operator, which was kind of
difficult for two reasons:

It changes the expression language. So far everything has been strictly
a subset of Javascript right hand side values, and assignment would add
the first left hand side and also the first statement.

This instantly raises problems like statement separation (; or similar)
and even async issues.

The strongest case for having assignment operations seem to be the
action expressions where we want to set property values of the objects
were are currently working with.

Here the async issues arise, because the action expressions do need to
produce a description of future operations to execute. We would need to
accept assignment operators as action chain parts to be either wrapped
automatically in functions or with a new construct.

In the end it seems far easier to just keep it like it is and rely
on a set action.