Component: Add class "iterative-context"
Add a "iterative-context" class to Datagrid. We will add this class to
all components creating context that is repeated multiple times.

DataGrid also receives a "!parentHasClass('iterative-context')" parent-rule
that excludes it from all so defined contexts.

The DataGrid data injection, like all current injection, works over simple
ids so the query would necessarily the same for for every iterated context
if it were in such a context.

If it should really make sense to have a repeated datagrid, you need to
provide the query result via expression in the corresponding prop.