Cleanup Attributes
Provide default constructor for attributes.

Change attribute values to be only String typed. The former possibility
of complex values was given up for XML-equivalence and easy editing.

Each attribute value is a string. If the string is wrapped in {} the
expression is parsed and the attribute either has an AST expression
value or is flagged as having an erroneous expression.

Catch the checked ParseException and wrap it in our generic
RuntimeException