uri Module: Refactor
We need to handle URL parameters in the incoming URL which we did not so
far. We include the "url" module for URL to extract the query params and
extend them with the params we're getting.

We add a third optional parameter "containsContextPath" which can be set
to true to signal that the URL already contains the context path to
avoid double adding without and ambiguities.