CodeEditor: Allow Props providing templates
Add support for the scenario where we are completing a prop value on a
component model and the value of that prop leads to different requirements on
the children of that model.

Concrete use case is the <Link.Param/>.

<Link text="Link" location="/foo/{id}"> requires a <Link.Param
name="id"/> to define the value.

So when a prop value is completed, we allow the completions returned to
contain a component model template which we insert as replacement for
the current model.

This has some implications on the way we handle cleaning up before
inserting the autocompleted value.

The completion cleanup works by selecting some amount of existing code
before the completion point, so far this was either the opening < or ".

Now the start point is the start of the current tag all the way to the
end of the current tag (not the completion point).

All this spreading over multiple lines.

We solve this by using the new "withPosition" feature of Tokens.toModel
giving us the location for the current tag and changing the internal
book keeping for completion start and end.