Attributes: Refactor
Move the attribute conversion out of the Attributes container into
AttributeValue.forValue methods with appropriate conversion into either
string literal or expression.

We offer methods for:

 * strings - ( expression if string starts with '{' and ends with '}')
 * numbers -> expr
 * booleans -> expr
 * maps -> expr
 * collection -> expr

All these types we could have expressed with a pure JSON model, but in
favor of clarity and XML-editability, we just use either string
attributes or convert the value to an expression which is converting it
to JSON with single quotes and wrapping it in "{ ... }".

Error Handling
--------------

We remove the last remains of the old just-bubble error handling.

Reminder: The new approach is that expressions which cannot be parsed
receive the special AttributeValueType.EXPRESSION_ERROR type. They keep
the broken expression as string value but return null as AST.