Refactor ExpressionEnvironment
The way we did the expression environment so far worked ok, but was a
bit clumsy and also not extendable.

This changes the way operations and identifiers are handled in
expression environment.

It draws the expression operations out of the expression enviroment, which
handles one expression language use case and encapsulates the system context
needed to evaluate the exression.

So far, the abstract base class handled all the method resolution for
us by inspecting all its extending classes and caching the introspection
result.

ExpressionService
-----------------

We now introduce a new service ExpressionService as main entry point.

public interface ExpressionService
{
    Object evaluate(Node node, ExpressionEnvironment env);
}

The method evaluate the given (partial) expression with the given
expression environment. The ExpressioService is configured with a set
of annotated operations classes which handle both operations
but also identifiers in better and more readable way.

Example
-------

@ExpressionOperations( environment = MyEnvironment.class)
public class Operations
{
    @Operation
    public Object myOp(ExpressionContext<MyEnvironment> ctx, String param)
    {
        ...
    }

    @Operation(context = MyContext.class)
    public Object myOp(
        ExpressionContext<MyEnvironment> ctx,
        MyContext myContext,
        String param)
    {
        ...
    }

    @Identifier
    public String ident(MyEnvironment.class)
    {
        ...
    }

}

Here we see an exemplary implementation of an operations class.

@ExpressionOperations
---------------------

The new @ExpressionOperations annotation marks operation classes and
receives the class of the ExpressionEnvironment implementation the
operations are declared for. It is a spring meta annotation, so it will
be autopicked up as spring bean if a component scan involving the
operation beans package is set up.

@Operation
----------

@Operation still marks the operation methods, but with a changed
signature.

The first paramteter has to be ExpressionContext which provides access
to the current environment the operation method is called in and the
ASTFunction of the operation for manual evaluation.

@Operation can now take the operation context as annotation parameter.
If such a paramter is given, the second parameter of the operation
method must be the context object.

Apart from these standard parameters, the user can also define
additional parameters which will be autoinjected by evaluating the
arguments to the ASTFunction and trying to cast them to the operations
method parameters.

If post-operation evaluation order of ASTFunction arguments is a problem
in your use case, you can still stick with manual evaluation of the
ASTFunction arguments.

@Identifier
-----------

There is a new way of declaratively defining identifiers with the
@Identifier annotation.

It marks a method producing the current identifier value which can only
the the environment as context.

The identifier name defaults to the method name but can be changed by
providing the name as annotation parameter.

Resolution Fallback
-------------------

Both resolveIdentifier and unsupportedOperation are kept and are called
whenever the ExpressionOperationService can not resolve an operation or
an identifier in the defined operation beans.

ReflectAsm
----------

ExpressionOperationService still uses reflection to analyze the
operation beans but uses ReflectASM to actually access the operations
and identifier methods. This seems like a good idea given that
reflection is quite a bit slower than the ReflectASM access (up to 800%)
and the expression evaluation is one big mess of interwoven operations
etc.

Using ReflectASM also cleans up the stackstraces quite a bit, so
runtime expression enviroment errors stack traces should look much
saner, too.