DataList: Add enums, rowCount
We add enum definitions for the enums used in the data list on the
server side and also provide the already existing separate total row
count used for paging in the data-list.

data-list.js module
-------------------

We change the data field of DataListCursor to "value" and provide a
valueOf and toString method that will return that value or the
stringification of that value. So for many purposes, the client can just
use cursors instead of the real values.

We make sure we store the enum registry and the row count
and provide a getRaw() method that returns the original JSON data of a
idataList without the runtime decorations.

We fix the path fed to onChange so that it includes the index of the
change.