Tokens: Provide component positions
Add second argument to toModel

	toModel(session, withPosition)

If set to true, withPosition causes all returned component models to
have a "pos" property defining their start and end position in the
current document.

We ignore this property on the server side to enable the client to send
these pos enhanced models without the server side caring.