RuntimeContext: Refactor
So far we have been using only the runtime application as larger
context to draw information from for other services. What we're missing
are a lot of the runtime information contained in our RuntimeContext
which is overly broad on the other hand.

Reduce RuntimeContext
---------------------

First we remove the request,response and modelmap from the runtime
context. We don't want and I think we also don't need all the
information available in theses objects or we only need them in
DefaultRuntimeApplication.route() which means we can pass it in
along side the runtime context and be ok.

For the rest of the application, we don't really need to access all
these fields. This also makes the context much more easily handable
in tests.

What we need in other parts of the application are the current mapping
parameter and HTTP parameters. We try to join both into one map and
provide that as variables in the runtime context.

Availability of runtime context
-------------------------------

The runtime context generally is available via threadlocal based
RuntimeContextHolder but we want to limit use of that for internal
purposes as much as we can.

We expose the runtime context to more sub components where it replaces
either the runtime application or the domain service as context.