Build: Shim react addons
The "react-addons-*" packages cause our bundle strategy of keeping react
itself shimed to fail cause these packages draw in a second copy of
react.js modules into our bundle, with require() paths we don't handle.

We combat this by switching back to the prebundled "react-with-addons"
bundle and then defining "react-addons-update" and
"react-addons-linked-state-mixin" as shims to the globals exposed from
that bundle.

Should we come up with a better way of making it work, should react
change their requires() in the future to use "react" internally, we can
revert this without having to change the uses of the addons.