Processes: Sketch model and loading
Create the initial model design for process based content and integrate
that ApplicationModel and load it via ModelCompositionService.

Mark views with the name of the process if they're used in a process.
Use that name to distinguish between views and process-contained views
in a viewInProcess() parent rule which allows components to express
that they are only valid in processes or only outside of processes.