Actions: Refactor
The data piping thing from the initial action implementation was in the
end a bad idea. It seemed cool and cute but it brings more problems then
it solves and it has no real benefit either.

It is important to have the ability to form complex action promise
chains but data piping just causes a lot of heaches on the server side
when trying to put that into any kind of meaningful data type. The
separation of action models as just statically describing the action
seemed like a good thing at first, but I now think it is rather bad.

Actions v2
==========

We keep the idea of Buttons (and maybe others) having an "action"
attribute which defines the actions in our expression language. We keep
the general action JSON format

	{
		"action" : "<action-name>",
		...
	}

But stop supporting arrays of actions and abandon autochaining the way
we did.

Action Expressions
------------------

We keep using expressions to express our actions but streamline this a
bit more by avoiding the long JSON format where it is possible in favor
of simple operations where that is possible.

For every action registered in the action service we create a default
operation with the same name taking in the action model without the name
identifier.

	myRegisteredAction({ param: 1 })

Action implementations can also implement a new interface

interface ActionCallGenerator
{
    void renderJsCode(ExpressionRenderer renderer, ASTFunction node);
}

The interface will be called from the ClientExpressionRenderer (that now
handles the action translation) to render the js code for a client-side
action invocation via the runtime view api.

Runtime View API Additions
--------------------------

The runtime view API contains some helper methods to deal with actions
and other common operations. The runtime view object itself is available
in the view as "_v" and offers the following methods

    _v.action(model[, name])

returns a promise for the execution of the given action. If the second
parameter is given, the "action" property of the model will be set to
that name, otherwise the model is assumed to already contain the action
property.

   _v.navigateTo(location[, params])

Navigates to the given internal location with the given location params.
The target must be an URI mapped to a view or a process, i.e. it has to
respond with the right JSON data etc. The method returns a promise that
resolves after the target view has been rendered.

  _v.update(id[,vars])

Updates component with the given id in the current view with new vars,
i.e. causes all queries of that component being refetched and the result
being rerendered.

  _v.uri(location, params)

Evaluate the given location with the given params and returns the
result.

 _v.observe(promise)

Observes the execution of the given promise. If the execution of the
promise takes longer than the throbber timeout, the throbber will be
automatically activated. Promise rejections will be logged as errors.

Action Chaining
---------------

The ClientExpressionHandler will compile the action compression to a
method containing an action call and will also provide default error
handling and action chaining.

An action expression like

	store(context).navigateTo('/')

will be compiled to something like

    function ()
    {
        return _v.observe(_v.action({
            action: "store",
            data: context
        }).then(function ()
        {
            return _v.navigateTo("/")
        }));
    }

Every action(-chain) will be wrapped in _v.observe() for auto-throbber
support and error logging.

Action Side-Effects
-------------------

Actions should be the primary driver of change in the application. This
follows Flux and other react patterns in creating a uni-directional flow
of data and a centralized point for these changes (in our case the
individual action implementations).

The actions cause changes inside the system and are for now just ack-ed.
We might automatically propagate some state changes back to the client
in the future (process scoped vars etc), but all in all the way to
actually see the changed state on the server, the action expression has
to cause either a navigation or an update.