DataList: Support object extraction
We add the ability to extract domain objects from DataList cursors where
it is applicable, i.e. where a cursor points to an actual nested domain
object or to a datalist row.

The method dataListCursor.getDomainObject( [type]) returns a valid domain
object with _type field for the given rowIndex.

If the type is not given, the only type used in the list will be used,
otherwise an error will be thrown.

For joined lists, the type must be given.

If the type is given for nested domain objects it is used to validate
the type of object the cursor is actually pointing to. The type is not
necessarry for nested object extraction though.