Domain: Support CRUD, refactor JOOQQueryExecutor
Add simple crud methods to our domain object service and convenience
methods on DomainObject itself routing over the internally stored domain
object service which should be normally be set at runtime.

The new CRUD methods use the non-DSL API of JOOQ (the objects being
built by the DSL API) which fits our purposes better and is much
clearer.

We change the JOOQQueryExecutor to use the non-DSL API, too and
generalize the naming strategy a bit to work in more contexts.