Add client side property converter
We already have a property converter on the server side to convert from
the java storage format to the JSON / Javascript format, but we still
need to be able to prepare values for display, like converting enums to
their (i18n) names and in the future also differences between JSON dates
and display dates etc.

We add a property-converter service module that can convert a value from
the display to the value format and back given the property type.