Model: Refactor
Rename enum model to EnumType. Make sure that the external models
connected to applicatio model are not just added by getting the
map and adding to it.

We wrap all maps with a read-only view and add dedicated add methods for
the external models.

This can also be used to initialize models as they're added to
application model.