Forms: Implement form components, calendar
Implement the base for a new form component system based on the
data-list features and the action features.

We use a High-Level component FormElement that wires up the actual input
element(s) via a value-link based API.

The FormElement HL-Component takes care of setting up the outer
bootstrap layout and handles the acquisition of a form element id and a
data cursor for the current field.

It uses a mixture of React context and Exceed contexts.

The Form Element creates an FormContext object that handles form errors
and form ids. It is transmitted as React Context:

        contextTypes: {
            formContext: React.PropTypes.instanceOf(FormContext)
        },

The data itself is transmitted via Exceed context features. The form
receives a data-list with one element via component query and keeps the
current state of that data-list as React component state.

FormElement creates DataList.Cursor objects based on the current name=""
/ path and

Internal Component Interface
----------------------------

The component wrapped by the high level component receives a standard
set of simplified props

    <FieldComponent
        id={ ... }
        valueLink={ ...  }
        className={ ... }
        validate={ ... }
        propertyType={ ... }
        onChange={ ... }
    />

 * id - Unique form element id. Must be rendered as HTML id attribute
   for the input (also label target)

 * valueLink - ValueLink/React link for the current converted value

 * className - HTML classes for the input. Might contain error classes

 * validate - function that can be called whenever the field value
              changes. Causes a validation of the field and updating of
              error classes and message

 * propertyType - the propertyType determined for the current cursor

 * onChange - callback function to call when the value really changes
              Will cause an immutable update to the underlying data-list

Example Form (Editor XML)
-------------------------

    <Form id="myForm"
        filter="{ id == param('id') }"
        horizontal="{ true }"
        type="{ Foo }">
        <StaticText name="Foo.id"/>
        <Field name="Foo.name"/>
        <Field name="Foo.created"/>
        <Field name="Foo.type"/>
        <Field name="Foo.num"/>
        <Toolbar>
            <Button ignoreErrors="{ true }"
                action="{ navigateTo('/') }"
                text="Cancel"/>
            <Button action="{ store(context).navigateTo('/') }"
                className="btn-primary"
                text="Save"/>
        </Toolbar>
    </Form>

The filter is used to reduce the query to a single element.

The design is to use <Field/> as general purpose element that adapts to
the property type.

For now it will mostly use a text input. It will switch to a select
element for enum properties and will add an button addon for Timestamp
and Date properties that pop ups an adaptation of the old OpenSAGA
calendar in React.

<StaticText/> shows an uneditable Text wrapped to align with form
elements.