Expressions: Add assignment and sequence
Exceed's expression language is working in a lot of contexts, some where
it works more natural then others.

For example the view expressions fit totally into the concept of
expressions with the expressions only being a description what type of
data to get from where, pure read access for everything but action
expressions.

Action expressions express a sequence of things which in the first
attempt was implemented as operation chaining. While this works it gets
very unwieldy for larger chains and it feels very unnatural.

Then there is the fact that having just an expression language we don't
have statements. Function expressions work well als replacement except
for assignment, where again it feels artificial and less readable than
just having an assignment operator.

This commit reintroduces the assignment operator which existed before
but which works now much better in conjunction with expression sequences
which we also introduce.

Expression Sequences
====================

Expression sequences are our new way to model action chains. The syntax
is

    a ; b ; c

Using good old semicolon as separator meaning, as-usual, "and then",
again with the difference that a,b,c are expressions, not statements.

  store() ; navigateTo('/')

Reads better then the chained variant and can easily be split into the
single actions without having much knowledge of the expression syntax
(aka being the client side).

Assignment Operator
===================

The assignment operator is a syntactically more pleasing replacement for
a set operator.

When we have a scoped context expression with the name 'myVar' we can
read it with

    property('myVar')

and then assign a value with

    property('myVar') = 1

We handle the complication of having to kinds of action syntaxes by
transforming an assigment operator into the appropriate set action early
when we parse the original expression for the component action prop /
decision / transition action.

The AssignmentReplacementVisitor converts

    property('v') = 1

into

    set({ type: 'PROPERTY', name: 'v', value: 1})

leaving us with one action format when we have to render / evaluate it.

ASTAssignment / ASTExpressionSequence
=====================================

The new AST types ASTAssignment and ASTExpressionSequence are invalid
for most expressions since they're only needed by action expressions in
either standalone view or transition / decision. ASTAssignment has no
runtime support at all and needs to transformed with aforementioned
AssignmentReplacementVisitor in any case.

ASTExpressionSequence is a sequence of multiple semicolon separated
expressions. The default ExpressionEnvironment behavior is to just
return a list of all values, very similar to the "[a,b,c]". The action
environment and view action renderer then override the defaults to
do client js action promise chaining or server-side action execution.